Move the regression helpers into the spec file since nothing else uses them.
